Changed way of getting package location.
When calling the 'what' function sometimes multiple instances of the package folder will be found which messes with the way the save and write functions of this class work. This is a quick fix! It would be better to use the "default paths" json interface to save the package location once.
Added the static method "getName" which will return only one path, which might be the correct location of the file.